Change Log Update (00.00.11)
* 2016-03-13 Long Break is Over
+ Started laying out Database Tables, Decided to stick with custom
version of SqliteWrapped.
* SqliteWrapped:
+ Makes lots of updates and have been rewriting SqliteWrapped to work a
little
easer with classes and help make some ORM related adjustments for
pulling data
in and out of classes easier.
+ Added Transactional inserts, updates, and creates, failures can no be
rolled back
and makes for quicker inserts of large data sets, and also creating
tables and indexes together.
* XRM:
+ Created a new Data Sub folder and initial data classes for Config and
Users.
Also created Config_Dao, and Users_Dao which will maintain the API for
reading and writing data.
+ Config file is setup as a serialization class that will read and write
to an XML file.
+ Users is setup to use the custom SQLite Wrappers API and updated
features and methods.
- Work has been a little start as my time is still limited.  I have to
rip out allot of the
flat file stuff I did and start writing interfaces for the data, then
later on write conversions
from the flat files to the new data layouts.
- Then i also need to decide what format the menu files will be in.
It's kind of silly to use a
database for menu files. So either XML, JSON, or something that is quick
and easy to edit and use.